Refactored Queue Forms to make consistent with other forms and between each other
